refactor(subagents): move nested runtime support behind in-process boundary
