+++ test/unit/subagents-structured-output-runtime.test.ts
@@ -5,6 +5,7 @@ import { join } from "node:path";
 import { describe, test } from "vitest";
 import type { ExtensionAPI, ToolDefinition } from "../../packages/coding-agent/src/index.js";
 import { rewriteSubagentPrompt } from "../../packages/subagents/src/runs/inprocess/prompt-behavior.js";
+import registerSubagentPromptRuntime from "../../packages/subagents/src/runs/inprocess/runtime-support/prompt-bridge.js";
 import {
 	cleanupStructuredOutputRuntime,
 	createStructuredOutputRuntime,
@@ -14,7 +15,6 @@ import {
 	STRUCTURED_OUTPUT_MISSING_ERROR,
 	STRUCTURED_OUTPUT_SCHEMA_ENV,
 } from "../../packages/subagents/src/runs/shared/structured-output.js";
-import registerSubagentPromptRuntime from "../../packages/subagents/src/runs/shared/subagent-prompt-runtime.js";
 
 const objectSchema = {
 	type: "object",
